# utils/swaggerApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "test API",
    "version": "v0.1"
  },
  "tags": [
    {
      "name": "Users API:",
      "description": "main path - https://yummy-rest-api.yurgo.fun/"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Get info of loginned user by token",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "code": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "fields update"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "UserName"
                        },
                        "email": {
                          "type": "string",
                          "example": "example@example.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Not authorized"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update existed user",
        "description": "Update name and/or avatar",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "example": {
                "name": "newName",
                "email": "Jimmy2updated@email.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "code": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "fields update"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "updatedUser": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Not authorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": [
              "false",
              "Set name of user"
            ]
          },
          "avatar": {
            "type": "string",
            "format": "binary"
          }
        }
      }
    }
  }
}